requirements.txt
gotrue==2.7.0
fastapi
uvicorn
pydantic
langchain
langchain_google_genai
aiohttp
python-telegram-bot==20.7
supabase
httpx